+++ includes/class-wc-order-refund-data-store-custom-table.php
@@ -6,6 +6,8 @@
  * @author  Liquid Web
  */
 
+use LiquidWeb\WooCommerceCustomOrdersTable\Concerns\UsesCustomTable;
+
 /**
  * Extend the WC_Order_Refund_Data_Store_CPT class, overloading methods that require database access in
  * order to use the new table.
@@ -13,57 +15,18 @@
  * This operates in a way similar to WC_Order_Data_Store_Custom_Table, but is for *refunds*.
  */
 class WC_Order_Refund_Data_Store_Custom_Table extends WC_Order_Refund_Data_Store_CPT {
+	use UsesCustomTable;
 
 	/**
-	 * Read refund data from the custom orders table.
+	 * Delete a refund from the database.
 	 *
-	 * If the refund does not yet exist, the plugin will attempt to migrate it automatically. This
-	 * behavior can be modified via the "wc_custom_order_table_automatic_migration" filter.
-	 *
-	 * @param WC_Order_Refund $refund      The refund object, passed by reference.
-	 * @param object          $post_object The post object.
+	 * @param WC_Order $refund The refund object, passed by reference.
+	 * @param array    $args  Additional arguments to pass to the delete method.
 	 */
-	protected function read_order_data( &$refund, $post_object ) {
-		$data = $this->get_order_data_from_table( $refund );
+	public function delete( &$refund, $args = array() ) {
+		add_action( 'woocommerce_delete_order_refund', [ $this, 'delete_row' ] );
 
-		if ( ! empty( $data ) ) {
-			$refund->set_props( $data );
-		} else {
-			/** This filter is defined in class-wc-order-data-store-custom-table.php. */
-			$migrate = apply_filters( 'wc_custom_order_table_automatic_migration', true );
-
-			if ( $migrate ) {
-				$this->populate_from_meta( $refund );
-			}
-		}
-	}
-
-	/**
-	 * Retrieve a single refund from the database.
-	 *
-	 * @global $wpdb
-	 *
-	 * @param WC_Order_Refund $refund The refund object.
-	 *
-	 * @return array The refund row, as an associative array.
-	 */
-	public function get_order_data_from_table( $refund ) {
-		global $wpdb;
-
-		$data = (array) $wpdb->get_row(
-			$wpdb->prepare(
-				'SELECT * FROM ' . esc_sql( wc_custom_order_table()->get_table_name() ) . ' WHERE order_id = %d LIMIT 1',
-				$refund->get_id()
-			),
-			ARRAY_A
-		); // WPCS: DB call OK.
-
-		// Expand anything that might need assistance.
-		if ( isset( $data['prices_include_tax'] ) ) {
-			$data['prices_include_tax'] = wc_string_to_bool( $data['prices_include_tax'] );
-		}
-
-		return $data;
+		parent::delete( $refund, $args );
 	}
 
 	/**
@@ -77,7 +40,7 @@ class WC_Order_Refund_Data_Store_Custom_Table extends WC_Order_Refund_Data_Store
 	protected function update_post_meta( &$refund ) {
 		global $wpdb;
 
-		$table       = wc_custom_order_table()->get_table_name();
+		$table       = self::get_custom_table_name();
 		$refund_data = array(
 			'order_id'           => $refund->get_id(),
 			'discount_total'     => $refund->get_discount_total( 'edit' ),
@@ -95,7 +58,7 @@ class WC_Order_Refund_Data_Store_Custom_Table extends WC_Order_Refund_Data_Store
 		);
 
 		// Insert or update the database record.
-		if ( ! wc_custom_order_table()->row_exists( $refund_data['order_id'] ) ) {
+		if ( ! $this->row_exists( $refund_data['order_id'] ) ) {
 			$inserted = $wpdb->insert( $table, $refund_data ); // WPCS: DB call OK.
 
 			if ( 1 !== $inserted ) {
@@ -109,43 +72,11 @@ class WC_Order_Refund_Data_Store_Custom_Table extends WC_Order_Refund_Data_Store
 				return;
 			}
 
-			$wpdb->update(
-				wc_custom_order_table()->get_table_name(),
-				$refund_data,
-				array( 'order_id' => (int) $refund->get_id() )
-			);
+			$wpdb->update( $table, $refund_data, [
+				'order_id' => $refund->get_id(),
+			] );
 		}
 
 		do_action( 'woocommerce_order_refund_object_updated_props', $refund, $refund_data );
-	}
-
-	/**
-	 * Populate the custom table row with post meta.
-	 *
-	 * @global $wpdb
-	 *
-	 * @param WC_Order_Refund $refund The refund object, passed by reference.
-	 * @param bool            $delete Optional. Whether or not the post meta should be deleted.
-	 *                                Default is false.
-	 *
-	 * @return WP_Error|null A WP_Error object if there was a problem populating the refund, or null
-	 *                       if there were no issues.
-	 */
-	public function populate_from_meta( &$refund, $delete = false ) {
-		global $wpdb;
-
-		$refund = WooCommerce_Custom_Orders_Table::populate_order_from_post_meta( $refund );
-
-		$this->update_post_meta( $refund );
-
-		if ( $wpdb->last_error ) {
-			return new WP_Error( 'woocommerce-custom-order-table-migration', $wpdb->last_error );
-		}
-
-		if ( true === $delete ) {
-			foreach ( WooCommerce_Custom_Orders_Table::get_postmeta_mapping() as $column => $meta_key ) {
-				delete_post_meta( $refund->get_id(), $meta_key );
-			}
-		}
 	}
 }