+++ tests/testcase.php
@@ -30,7 +30,7 @@ class TestCase extends WC_Unit_Test_Case {
 		WooCommerce_Custom_Orders_Table_Install::activate();
 
 		$wpdb->suppress_errors( false );
-		$wpdb->query( 'DELETE FROM ' . esc_sql( wc_custom_order_table()->get_table_name() ) );
+		$wpdb->query( 'DELETE FROM ' . esc_sql( wc_custom_order_table()->get_orders_table_name() ) );
 	}
 
 	/**
@@ -68,29 +68,7 @@ class TestCase extends WC_Unit_Test_Case {
 	}
 
 	/**
-	 * Given an array of IDs, see how many of those IDs exist in the table.
-	 *
-	 * @global $wpdb
-	 *
-	 * @param array $order_ids An array of order IDs to look for.
-	 *
-	 * @return int The number of matches found in the database.
-	 */
-	protected function count_orders_in_table_with_ids( $order_ids = array() ) {
-		global $wpdb;
-
-		if ( empty( $order_ids ) ) {
-			return 0;
-		}
-
-		return (int) $wpdb->get_var( $wpdb->prepare(
-			'SELECT COUNT(order_id) FROM ' . esc_sql( wc_custom_order_table()->get_table_name() ) . '
-			WHERE order_id IN (' . implode( ', ', array_fill( 0, count( (array) $order_ids ), '%d' ) ) . ')',
-		$order_ids ) );
-	}
-
-	/**
-	 * Retrieve a single row from the Orders table.
+	 * Retrieve a single row from the orders table.
 	 *
 	 * @global $wpdb
 	 *
@@ -102,7 +80,7 @@ class TestCase extends WC_Unit_Test_Case {
 		global $wpdb;
 
 		return $wpdb->get_row( $wpdb->prepare(
-			'SELECT * FROM ' . esc_sql( wc_custom_order_table()->get_table_name() ) . ' WHERE order_id = %d',
+			'SELECT * FROM ' . esc_sql( wc_custom_order_table()->get_orders_table_name() ) . ' WHERE order_id = %d',
 			$order_id
 		), ARRAY_A );
 	}